+++ src/manage.py
@@ -1,31 +1,27 @@
 import unittest
 import os
 import coverage
-
 from flask_script import Manager
-
-
 from app import app
 
-
 manager = Manager(app)
-
-
 
 @manager.command
 def test():
     """Runs the unit tests without coverage."""
-    tests = unittest.TestLoader().discover('.')
+    tests = unittest.TestLoader().discover('_test')
     unittest.TextTestRunner(verbosity=2).run(tests)
-
 
 @manager.command
 def cov():
     """Runs the unit tests with coverage."""
-    cov = coverage.coverage(branch=True, omit=['*/site-packages/*', 'test.py'])
+    cov = coverage.coverage(branch=True, omit=[
+                                            '*/site-packages/*',
+                                            '*test*.py'
+                                            ])
     cov.start()
-    tests = unittest.TestLoader().discover('.')
-    unittest.TextTestRunner(verbosity=2).run(tests)
+    tests = unittest.TestLoader().discover('_test')
+    unittest.TextTestRunner(verbosity=5).run(tests)
     cov.stop()
     cov.save()
     print 'Coverage Summary:'
