+++ src/server/data_access/user_data_access.py
@@ -6,7 +6,7 @@ class UserDataAccess:
         self.users = users
 
     def authorize(self, requestForm):
-        user, status, message = UserDataAccess.init_output()
+        user, status, message = {}, False, ""
         form = LoginForm(requestForm)
         if form.validate():
             login_user = self.users.find_one({
@@ -14,8 +14,7 @@ class UserDataAccess:
                 })
 
             if login_user and UserDataAccess.check_ps(login_user, form.password.data):
-                    status = True
-                    message = 'Login successful!'
+                    status, message = True, 'Login successful!'
                     user = UserDataAccess.return_user(login_user)
             else:
                 message = 'The username and password does not match!'
@@ -30,9 +29,8 @@ class UserDataAccess:
         == login_user["password"].encode('utf-8')
 
     def register(self, requestForm):
-        user, status, message = UserDataAccess.init_output()
+        user, status, message = {}, False, ""
         form = RegisterForm(requestForm)
-
         if form.validate():
             is_unique, message = self.__is_unique(form.username.data, form.email.data)
             if not is_unique:
@@ -64,17 +62,12 @@ class UserDataAccess:
         message = 'You can use this username and email!'
         status = True
         if self.users.find_one({'username' : username}):
-            status = False
-            message = 'The username has been taken!'
+            status, message = False, 'The username has been taken!'
         if status:
             if self.users.find_one({'email' : email}):
                 status = False
                 message = 'The email has been taken!'
         return status, message
-
-    @staticmethod
-    def init_output():
-        return {}, False, ""
 
     @staticmethod
     def return_user(user_info):
@@ -101,7 +94,7 @@ class UserDataAccess:
         return True, 'You can use this email!'
 
     def update_profile(self, requestForm):
-        user, status, message = UserDataAccess.init_output()
+        status, message = False, ""
         form = UpdateProfileForm(requestForm)
         if form.validate():
             status, message = self.__is_your_email_unique(form.email.data)
@@ -115,7 +108,7 @@ class UserDataAccess:
         return UserDataAccess.return_output(status, message, {})
 
     def change_password(self, requestForm, username):
-        user, status, message = UserDataAccess.init_output()
+        status, message = False, ""
         form = ChangePasswordForm(requestForm)
         if form.validate():
             login_user = self.users.find_one({
@@ -128,8 +121,7 @@ class UserDataAccess:
                         bcrypt.hashpw(form.newpassword.data.encode('utf-8'), bcrypt.gensalt())
                     }}
                 )
-                status = True
-                message = 'Your password has been changed!'
+                status, message  = True, 'Your password has been changed!'
             else:
                 message = 'The old password is NOT correct!'
         else:
@@ -137,7 +129,7 @@ class UserDataAccess:
         return UserDataAccess.return_output(status, message, {})
 
     def delete(self, requestForm):
-        user, status, message = UserDataAccess.init_output()
+        status, message = False, ""
         form = DeleteForm(requestForm)
         if form.validate():
             login_user = self.users.find_one({
@@ -148,8 +140,7 @@ class UserDataAccess:
                 self.users.delete_one(
                     { 'username': form.username.data}
                 )
-                status = True
-                message = 'Your account has been deleted!'
+                status, message = True, 'Your account has been deleted!'
         else:
             message = 'Missing info!'
         return UserDataAccess.return_output(status, message, {})
